Further abstracting main
