+++ crates/cli/src/main.rs
@@ -6,14 +6,14 @@ use std::error::Error;
 
 use bindings::uniswap_v3_pool;
 use clap::{CommandFactory, Parser, Subcommand};
-use ethers::types::U256;
 use eyre::Result;
-use on_chain::monitor::{EventMonitor, HistoricalMonitor};
+use on_chain::monitor::EventMonitor;
 
 mod config;
 mod gbm;
 mod ou;
 mod sim;
+mod backtest_data;
 
 #[derive(Parser)]
 #[command(name = "Arbiter")]
@@ -74,6 +74,9 @@ enum Commands {
         /// Path to config.toml containing simulation parameterization (optional)
         #[arg(short, long, default_value = "./crates/cli/src/config.toml", num_args = 0..=1)]
         config: String,
+        /// Path to csv file containing price data
+        #[arg(short = 'f', long, required = true)]
+        file_path: String,
     },
 }
 
@@ -109,38 +112,18 @@ async fn main() -> Result<(), Box<dyn Error>> {
         }
 
         Some(Commands::ExportSwapRange {
-            config: _,
+            config,
             start_block,
             end_block,
             address,
         }) => {
-            let range = *start_block..*end_block;
-            let step = 100_u64; // doing this so we don't hit rpc limits
-            let contract_address = address;
-            let historical_monitor =
-                HistoricalMonitor::new(on_chain::monitor::utils::RpcTypes::Mainnet).await;
-            let contract_abi = uniswap_v3_pool::UNISWAPV3POOL_ABI.to_owned();
-            let mut pricedata: Vec<U256> = Vec::new();
-            for block in range.step_by(step as usize) {
-                let sqrtpricex96 = historical_monitor
-                    .historical_monitor(contract_address, contract_abi.clone(), block, block + step)
-                    .await;
-                let sqrtpricex96 = sqrtpricex96.unwrap();
-                pricedata.extend(sqrtpricex96)
-            }
-
-            historical_monitor
-                .save_price_to_csv(&pricedata, "price.csv")
-                .unwrap();
+            // Export swap price data for a given block range
+            backtest_data::save_backtest_data(config, start_block, end_block, address).await;
         }
 
-        Some(Commands::Importbacktest { config: _ }) => {
-            let price_data =
-                simulate::price_simulation::import_price_from_csv("price_data.csv").unwrap();
-            let price_ref = &price_data;
-            let _ = price_ref;
-
-            println!("{:?}", price_ref);
+        Some(Commands::Importbacktest { config, file_path}) => {
+            // Import swap price data from a csv file
+            backtest_data::load_backtest_data(config, file_path).await;
         }
 
         None => {
